+++ Back-end/MammaMia/src/main/java/com/pizzeria/MammaMia/Dto/ProductDTO.java
@@ -1,26 +1,20 @@
 package com.pizzeria.MammaMia.Dto;
 
 import com.pizzeria.MammaMia.Entity.Flavor;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProductDTO {
-    private Long id;
+    private Integer id;
     private String productName;
     private String productDescription;
-    private Float price;
-
-//    private int productFlavor;
+    private BigDecimal price;
     private int quantity;
     private Flavor flavor;
-
-
-
-
 }
