+++ Back-end/MammaMia/src/main/java/com/pizzeria/MammaMia/Entity/Client.java
@@ -11,7 +11,7 @@ public class Client {
     @Id
     @Getter @Setter
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     @OneToOne
     @Getter @Setter
     @JoinColumn(name = "user_id")
@@ -28,7 +28,7 @@ public class Client {
     private String phone;
     public Client() {
     }
-    public Client(Long id, RegisterUser registerUser, Address address, String cpf, String name, String phone) {
+    public Client(Integer id, RegisterUser registerUser, Address address, String cpf, String name, String phone) {
         this.id = id;
         this.registerUser = registerUser;
         this.address = address;
