+++ Back-end/MammaMia/src/main/java/com/pizzeria/MammaMia/Entity/DeliveryPeople.java
@@ -13,14 +13,14 @@ import lombok.NoArgsConstructor;
 public class DeliveryPeople {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "employ_id")
     private Employ employ;
     private String cpf;
     private String name;
     private String phone;
-    public DeliveryPeople(Long id, Employ employ, String cpf, String name, String phone) {
+    public DeliveryPeople(Integer id, Employ employ, String cpf, String name, String phone) {
         this.id = id;
         this.employ = employ;
         this.cpf = cpf;
